# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CompletionRequest": {
        "additionalProperties": false,
        "properties": {
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          }
        },
        "title": "CompletionRequest",
        "type": "object"
      },
...
      "EngagementPath": {
        "enum": [
          "first_time",
          "reactivation"
        ],
        "title": "EngagementPath",
        "type": "string"
      },
...
      "EngagementRead-Output": {
        "properties": {
          "billable_start_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billable Start At"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "confirmed_at": {
            "format": "date-time",
            "title": "Confirmed At",
            "type": "string"
          },
          "contract_sent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract Sent At"
          },
          "contract_terms": {
            "$ref": "#/components/schemas/ContractTerms"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "decline": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeclineRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeedbackRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "path": {
            "$ref": "#/components/schemas/EngagementPath"
          },
          "prefilled_from_engagement_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefilled From Engagement Id"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "rate": {
            "title": "Rate",
            "type": "string"
          },
          "signed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed At"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/EngagementStatus"
          },
          "stuck": {
            "title": "Stuck",
            "type": "boolean"
          },
          "work_mode": {
            "$ref": "#/components/schemas/WorkMode"
          },
          "worker_id": {
            "format": "uuid",
            "title": "Worker Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "worker_id",
          "project_id",
          "path",
          "status",
          "start_date",
          "end_date",
          "rate",
          "currency",
          "work_mode",
          "location",
          "contract_terms",
          "prefilled_from_engagement_id",
          "confirmed_at",
          "contract_sent_at",
          "signed_at",
          "billable_start_at",
          "completed_at",
          "stuck",
          "feedback"
        ],
        "title": "EngagementRead",
        "type": "object"
      },
...
      "FeedbackCreate": {
        "additionalProperties": false,
        "properties": {
          "free_text": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Text"
          },
          "skill_ids_demonstrated": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 20,
            "title": "Skill Ids Demonstrated",
            "type": "array"
          },
          "structured_answers": {
            "$ref": "#/components/schemas/StructuredAnswers"
          }
        },
        "required": [
          "structured_answers"
        ],
        "title": "FeedbackCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProjectCreate": {
        "additionalProperties": false,
        "properties": {
          "client_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Name"
          },
          "data_region": {
            "pattern": "^[A-Z]{2,8}$",
            "title": "Data Region",
            "type": "string"
          },
          "ends_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends On"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "required_skill_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 50,
            "title": "Required Skill Ids",
            "type": "array"
          },
          "starts_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts On"
          }
        },
        "required": [
          "name",
          "data_region"
        ],
        "title": "ProjectCreate",
        "type": "object"
      },
...
      "WorkMode": {
        "enum": [
          "remote",
          "onsite",
          "hybrid"
        ],
        "title": "WorkMode",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Bonarda Works API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/engagements/{engagement_id}/complete": {
      "post": {
        "operationId": "complete_engagement_api_v1_engagements__engagement_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "engagement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Engagement Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementRead-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Complete Engagement",
        "tags": [
          "engagements"
        ]
      }
    },
    "/api/v1/engagements/{engagement_id}/contract/retry": {
      "post": {
        "operationId": "retry_contract_api_v1_engagements__engagement_id__contract_retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "engagement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Engagement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementRead-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retry Contract",
        "tags": [
          "engagements"
        ]
      }
    },
    "/api/v1/engagements/{engagement_id}/feedback": {
      "post": {
        "operationId": "submit_feedback_api_v1_engagements__engagement_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "engagement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Engagement Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementRead-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Feedback",
        "tags": [
          "engagements"
        ]
      }
    },
...
    "/api/v1/projects": {
      "get": {
        "operationId": "list_projects_api_v1_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectRead"
                  },
                  "title": "Response List Projects Api V1 Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Projects",
        "tags": [
          "engagements"
        ]
      },
      "post": {
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Project",
        "tags": [
          "engagements"
        ]
      }
    },
    "/api/v1/projects/{project_id}": {
      "get": {
        "operationId": "read_project_api_v1_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Project",
        "tags": [
          "engagements"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/close": {
      "post": {
        "operationId": "close_project_api_v1_projects__project_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Close Project",
        "tags": [
          "engagements"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/staff": {
      "put": {
        "operationId": "set_project_staff_api_v1_projects__project_id__staff_put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffAssignment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Project Staff",
        "tags": [
          "engagements"
        ]
      }
    },
...
    "/api/v1/webhooks/esign": {
      "post": {
        "operationId": "esign_webhook_api_v1_webhooks_esign_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-bonarda-timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Bonarda-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "x-bonarda-signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Bonarda-Signature"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Esign Webhook",
        "tags": [
          "engagements"
        ]
      }
    },
...
    "/api/v1/workers/me/engagements/{engagement_id}/decline": {
      "post": {
        "description": "The worker declines an offer they have not signed (offer-decline spec \u00a74).",
        "operationId": "decline_engagement_api_v1_workers_me_engagements__engagement_id__decline_post",
        "parameters": [
          {
            "in": "path",
            "name": "engagement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Engagement Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeclineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementRead-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decline Engagement",
        "tags": [
          "engagements"
        ]
      }
    },
...
    "/api/v1/workers/{worker_id}/engagements": {
      "get": {
        "operationId": "list_worker_engagements_api_v1_workers__worker_id__engagements_get",
        "parameters": [
          {
            "in": "path",
            "name": "worker_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Worker Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EngagementRead-Output"
                  },
                  "title": "Response List Worker Engagements Api V1 Workers  Worker Id  Engagements Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Worker Engagements",
        "tags": [
          "engagements"
        ]
      },
      "post": {
        "operationId": "create_first_time_engagement_api_v1_workers__worker_id__engagements_post",
        "parameters": [
          {
            "in": "path",
            "name": "worker_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Worker Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngagementCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementRead-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create First Time Engagement",
        "tags": [
          "engagements"
        ]
      }
    },
    "/api/v1/workers/{worker_id}/reactivation-prefill": {
      "get": {
        "operationId": "reactivation_prefill_api_v1_workers__worker_id__reactivation_prefill_get",
        "parameters": [
          {
            "in": "path",
            "name": "worker_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Worker Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReactivationPrefill"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reactivation Prefill",
        "tags": [
          "engagements"
        ]
      }
    },
    "/api/v1/workers/{worker_id}/reactivations": {
      "post": {
        "operationId": "reactivate_worker_api_v1_workers__worker_id__reactivations_post",
        "parameters": [
          {
            "in": "path",
            "name": "worker_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Worker Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactivationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementRead-Input"
                }
              }
            },
            "description": "Replayed Idempotency-Key"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementRead-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reactivate Worker",
        "tags": [
          "engagements"
        ]
      }
    },
...
  }
}